Time: 1 ms (43.06%), Space: 17.8 MB (66.73%) - LeetHub

+++ 1903-largest-odd-number-in-string/1903-largest-odd-number-in-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string largestOddNumber(string num) {
-        for(int i = (int)num.size()-1 ; i>=0 ; --i){
+        for(int i =num.size()-1 ; i>=0 ; --i){
             int d = num[i]-'0';
             if(d%2==1){
                 return num.substr(0,i+1);
